refactor(routing): move dynamic routing related code changes to separate files
